Automation/requirements.txt
filelock==3.8.0
jsonpath-rw==1.4.0
loguru==0.7.0
pytest==7.3.1
python-dateutil==2.8.2
PyYAML==6.0
Requests==2.31.0
allure-pytest
faker
rsa==4.9
